Using context from IOHIDManager callbacks to call Obj-c methods
Really impressed with this solution I gathered off stack overflow. The key is that you want to
call an objective-c method, but this interface only takes a function pointer. You can simply
wrap your current state in the context (as void *) and then, in the C function, you unwrap your
class and call the method on it. Quite brilliant.

http://stackoverflow.com/questions/1772349/does-this-function-iohidmanagerregisterdevicematchingcallback-operate-under-the