All plumbed up and working
At this point, the project is a NSSystemBar menu that has some actions.
You can select individual ones, but they do not show which is actually
active. When the USB panic button is depressed, it will call the action
for whichever menu item was last selected. Right now it's only logging,
so it's not terribly impressive, but I'm happy with how this is
progressing now.